+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Next release
 
+
 - feat: added chain-id to the GenesisConfig in pallet-starknet
+- chore(rpc): clean trace api
+- feat(rpc): added state diff real value in trace api
 - chore: update cairo-vm commit and update gas per op
 - refactor(rpc): use single arc instance of starknet rpc
 - build: remove patch on `ring-vrf` dependecy